Move program gen to program files
